Delete multiple files to maintain max percent
The max percent command line option (-M) only results in the oldest log file in the log directory being deleted when the max disk usage percentage is reached. This is problematic since variations in logged traffic over time can result in some PCAP files being smaller than others, and this causes the max percentage to be exceeded at times. This fix will delete as many of the oldest PCAP files in the log directory as is needed to maintain the specified max percent value.